added dto and completed the base code, can get users from db, next step  is to be able to add transactions

+++ src/main/java/com/bank/WalletApp/persistence/entity/Transaction.java
@@ -27,7 +27,7 @@ public class Transaction {
 
     @NonNull
     @Column(name = "Amount")
-    private Double amount;
+    private Long amount;
 
     @Temporal(TemporalType.TIMESTAMP)
     private Date date;
